requirements.txt
flask
flask-cors
requests
gunicorn
python-dotenv
google-generativeai
gTTS